feature[TW19944]: Continuation IDs - Replace SQL getInt() for artifact ids with getLong()

+++ plugins/org.eclipse.osee.ats.rest/src/org/eclipse/osee/ats/rest/internal/util/health/AtsHealthCheckOperation.java
@@ -493,7 +493,7 @@ public class AtsHealthCheckOperation {
       try {
          chStmt.runPreparedQuery(AtsHealthQueries.getWorkItemsInCurrentStateType(atsApi, StateType.Working));
          while (chStmt.next()) {
-            artIds.add(Long.valueOf(chStmt.getInt(1)));
+            artIds.add(chStmt.getLong(1));
          }
       } finally {
          chStmt.close();
